add generator ICO picture

+++ admin/core/CmsLite.php
@@ -437,16 +437,22 @@ class CmsLite
                 return $this->prepare($this->_result);
             case 'save-ico':
                 $fileName = basename($data['name']);
-                $source = Helper::normalizePath(Loader::$rootDir . DIRECTORY_SEPARATOR . 'img'. DIRECTORY_SEPARATOR.$fileName);
-                $destination = Helper::normalizePath(Loader::$rootDir . DIRECTORY_SEPARATOR . 'img'. DIRECTORY_SEPARATOR .'example.ico');
-                move_uploaded_file($data['tmp_name'], $source);
-                $ico_lib = new ICOGenerator($source, [[32, 32]]);
-                $ico_lib->saveICO($destination);
-                unlink($source);
-                $this->_result['success'] = true;
-                $this->_result['message'] = 'Favicon успешно установлен!';
-                $this->_result['type'] = 'success';
-                $this->_result['link-ico'] = Url::to('img/example.ico');
+                if(!empty($fileName)){
+                    $source = Helper::normalizePath(Loader::$rootDir . DIRECTORY_SEPARATOR . 'img'. DIRECTORY_SEPARATOR.$fileName);
+                    $destination = Helper::normalizePath(Loader::$rootDir . DIRECTORY_SEPARATOR . 'img'. DIRECTORY_SEPARATOR .'example.ico');
+                    move_uploaded_file($data['tmp_name'], $source);
+                    $ico_lib = new ICOGenerator($source, [[32, 32]]);
+                    $ico_lib->saveICO($destination);
+                    unlink($source);
+                    $this->_result['success'] = true;
+                    $this->_result['message'] = 'Favicon успешно установлен!';
+                    $this->_result['type'] = 'success';
+                    $this->_result['link-ico'] = Url::to('img/example.ico');
+                } else {
+                    $this->_result['success'] = false;
+                    $this->_result['message'] = 'Нет данных для сохранения!';
+                    $this->_result['type'] = 'warn';
+                }
                 return $this->prepare($this->_result);
             default:
                 //TODO
